fix: add to decimal material colum tkdn & fixing all migrate

+++ database/migrations/2025_07_14_045543_create_equipment_table.php
@@ -11,23 +11,25 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('equipment', function (Blueprint $table) {
-            $table->charset = 'utf8mb4';
-            $table->collation = 'utf8mb4_unicode_ci';
-            
-            $table->ulid('id')->primary();
-            $table->string('code')->unique();
-            $table->string('name');
-            $table->ulid('category_id')->nullable();
-            $table->decimal('tkdn', 5, 2)->nullable(); // TKDN as percentage, e.g. 87.50
-            $table->integer('period')->comment('Satuan Hari');
-            $table->unsignedBigInteger('price');
-            $table->string('description')->nullable();
-            $table->string('location')->nullable();
-            $table->timestamps();
-            
-            $table->foreign('category_id')->references('id')->on('categories')->onDelete('set null');
-        });
+        if (!Schema::hasTable('equipment')) {
+            Schema::create('equipment', function (Blueprint $table) {
+                $table->charset = 'utf8mb4';
+                $table->collation = 'utf8mb4_unicode_ci';
+                
+                $table->ulid('id')->primary();
+                $table->string('code')->unique();
+                $table->string('name');
+                $table->ulid('category_id')->nullable();
+                $table->decimal('tkdn', 5, 2)->nullable(); // TKDN as percentage, e.g. 87.50
+                $table->integer('period')->comment('Satuan Hari');
+                $table->unsignedBigInteger('price');
+                $table->string('description')->nullable();
+                $table->string('location')->nullable();
+                $table->timestamps();
+                
+                $table->foreign('category_id')->references('id')->on('categories')->onDelete('set null');
+            });
+        }
     }
 
     /**
